+++ lib/src/components/components.dart
@@ -35,3 +35,4 @@ export 'flowbite_progress_bar.dart';
 export 'flowbite_spinner.dart';
 export 'flowbite_toast.dart';
 export 'flowbite_toggle_input.dart';
+export 'flowbite_tooltip.dart';