+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `FlowbiteBreadcrumb` component.
 - Add `FlowbiteToast` component.
 - Add `FlowbiteProgressBar` component.
+- Add `FlowbiteTooltip` component.
 
 ## [0.0.5] - 11-09-2026
 ### Added